expander updated to include 0 and updated to ignore non existing env keys
